requirements.txt
async-timeout==5.0.1
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==3.1.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
python-dotenv==1.0.1
redis==5.2.0
requests==2.32.3
spotipy==2.24.0
urllib3==1.26.6
Werkzeug==3.1.3
zipp==3.21.0
gunicorn==21.2.0
Flask-Cors==4.0.0
